requirements.txt
# CourseMate dependencies

customtkinter>=5.0.0
requests>=2.31.0
Pillow>=9.0.0
